LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/form-control-infrastructure/form-expected.txt


PASS button.form
PASS fieldset.form
PASS input.form
PASS keygen.form
PASS object.form
PASS output.form
PASS select.form
PASS textarea.form
PASS label.form
PASS label-form.form
PASS label-form-form2.form
PASS label-with-control.form
PASS label-for.form
FAIL label-with-progress.form assert_equals: Sanity check: label.control.form expected (object) null but got (undefined) undefined
FAIL label-with-meter.form assert_equals: Sanity check: label.control.form expected (object) null but got (undefined) undefined
PASS label-for-control-form-in-form.form
PASS label-for-control-form.form
PASS label-in-table.form
PASS label-in-table-with-control.form
PASS label-in-table-for.form
button

fieldset




object

output





label

label-form

label-form-form2

label-with-control

label-for

label-with-progress

label-with-meter

 label-for-control-form-in-form

 label-for-control-form

label-in-table	label-in-table 	label-in-table-for	
